third_party/WebKit/LayoutTests/fast/domurl/urlsearchparams.html

<!doctype html>
<html>
<head>
<meta charset="utf8">
<link rel="help" href="https://url.spec.whatwg.org/#interface-urlsearchparams">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
    assert_true('URLSearchParams' in window);
    assert_true('toString' in URLSearchParams.prototype);
    assert_true('append' in URLSearchParams.prototype);
    assert_true('delete' in URLSearchParams.prototype);
    assert_true('get' in URLSearchParams.prototype);
    assert_true('getAll' in URLSearchParams.prototype);
    assert_true('has' in URLSearchParams.prototype);
    assert_true('set' in URLSearchParams.prototype);
}, 'URLSearchParams interface');
</script>
</head>
</html>
